+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.3.4 (2024--)
 
 - `objDeleteKeyByValue`: Add `objDeleteKeyByValue` method
+- `objUpdate`: Add `objUpdate` method
 
 ## 1.3.3 (2024-03-05)